move ErrorHandlingTextField to TextField file
