+++ src/main/typescript/lib/formComponents/ContributorFieldArray.tsx
@@ -16,8 +16,7 @@
 import * as React from "react"
 import asFieldArray, { InnerComponentProps } from "./FieldArrayHOC"
 import { Field } from "redux-form"
-import TextField from "./TextField"
-import ErrorHandlingTextField from "./ErrorHandlingTextField"
+import TextField, { ErrorHandlingTextField } from "./TextField"
 import { DropdownFieldInput } from "./DropDownField"
 import { DropdownListEntry } from "../../model/DropdownLists"
 import RemoveButton from "./RemoveButton"
