+++ src/main/typescript/lib/formComponents/SchemedTextFieldArray.tsx
@@ -15,7 +15,7 @@
  */
 import * as React from "react"
 import { Field } from "redux-form"
-import TextField from "./TextField"
+import { ErrorHandlingTextField } from "./TextField"
 import { DropdownFieldInput } from "./DropDownField"
 import { DropdownListEntry } from "../../model/DropdownLists"
 import asFieldArray, { InnerComponentProps } from "./FieldArrayHOC"
@@ -39,7 +39,7 @@ const SchemedTextField = ({ names, label, schemeValues, withEmptyDefault }: Sche
             <Field name={names[1]}
                    label="Value"
                    placeholder={label}
-                   component={TextField}/>
+                   component={ErrorHandlingTextField}/>
         </div>
     </div>
 )
